Rename window module to application
This removed the former application implementation.
