+++ examples/custom_titlebar.rs
@@ -4,12 +4,12 @@
 
 fn main() -> wry::Result<()> {
   use wry::{
-    webview::WebViewBuilder,
-    window::{
+    application::{
       event::{Event, WindowEvent},
       event_loop::{ControlFlow, EventLoop},
       window::WindowBuilder,
     },
+    webview::WebViewBuilder,
     RpcRequest,
   };
