+++ examples/fullscreen.rs
@@ -4,12 +4,12 @@
 
 fn main() -> wry::Result<()> {
   use wry::{
-    webview::WebViewBuilder,
-    window::{
+    application::{
       event::{Event, WindowEvent},
       event_loop::{ControlFlow, EventLoop},
       window::{Fullscreen, WindowBuilder},
     },
+    webview::WebViewBuilder,
   };
 
   let event_loop = EventLoop::new();
