+++ src/application/icon.rs
