--- src/window/mod.rs
@@ -1,11 +1,0 @@
-// Copyright 2019-2021 Tauri Programme within The Commons Conservancy
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-License-Identifier: MIT
-
-pub use winit::dpi;
-pub mod error;
-pub mod event;
-pub mod event_loop;
-mod icon;
-pub mod monitor;
-pub mod window;
